clippy: clear the driver and test-suite lint backlog
Modernize the Vulkan driver (c"" literals, plain use over pub(self),
derived Default on FFI structs, is_multiple_of, dropped redundant casts)
and quiet the scattered test-file lints (arbitrary 3.14 constants that
tripped approx_constant, index-driven loops), with documented allows
where a lint fights readability (shared Arc over an FFI handle, the large
ICB command variant, double-duty loop indices).

+++ src/driver/vulkan/device_impl.rs
@@ -1487,7 +1487,7 @@ impl GpuDevice for VulkanDevice {
 
         // Allocate a buffer to track the array binding. Full implementation would create
         // a descriptor set with variable descriptor count.
-        let size = (textures.len().max(1) * 8) as usize;
+        let size = textures.len().max(1) * 8;
         let handle = self.field_alloc_impl(size, FieldUsage::READ.union(FieldUsage::TRANSFER))?;
         Ok(handle)
     }
@@ -1504,7 +1504,7 @@ impl GpuDevice for VulkanDevice {
         }
         drop(buf_map);
 
-        let size = (buffers.len().max(1) * 8) as usize;
+        let size = buffers.len().max(1) * 8;
         let handle = self.field_alloc_impl(size, FieldUsage::READ.union(FieldUsage::TRANSFER))?;
         Ok(handle)
     }
